Add polyfill
