+++ libs/blocks/caas/caas.js
@@ -1,6 +1,17 @@
-/* eslint-disable compat/compat */
-import { initCaas, loadCaasFiles, loadStrings, fgHeaderValue } from './utils.js';
-import { parseEncodedConfig, createIntersectionObserver, getMetadata, getConfig, b64ToUtf8 } from '../../utils/utils.js';
+import {
+  decodeCompressedString,
+  fgHeaderValue,
+  initCaas,
+  loadCaasFiles,
+  loadStrings,
+} from './utils.js';
+import {
+  b64ToUtf8,
+  createIntersectionObserver,
+  getConfig,
+  getMetadata,
+  parseEncodedConfig,
+} from '../../utils/utils.js';
 
 const ROOT_MARGIN = 1000;
 const P_CAAS_AIO = b64ToUtf8('MTQyNTctY2hpbWVyYS5hZG9iZWlvcnVudGltZS5uZXQvYXBpL3YxL3dlYi9jaGltZXJhLTAuMC4xL2NvbGxlY3Rpb24=');
@@ -13,32 +24,6 @@ const getCaasStrings = (placeholderUrl) => new Promise((resolve) => {
   }
   resolve({});
 });
-
-const decodeCompressedString = async (txt) => {
-  const b64decode = (str) => {
-    const binaryStr = window.atob(str);
-    const bytes = new Uint8Array(new ArrayBuffer(binaryStr.length));
-    binaryStr?.split('')
-      .forEach((c, i) => (bytes[i] = c.charCodeAt(0)));
-    return bytes;
-  };
-
-  const b64toStream = (b64) => new Blob([b64decode(b64)], { type: 'text/plain' }).stream();
-
-  const decompressStream = async (stream) => new Response(
-    // eslint-disable-next-line no-undef
-    stream.pipeThrough(new DecompressionStream('gzip')),
-  );
-
-  const responseToJSON = async (response) => {
-    const blob = await response.blob();
-    return JSON.parse(await blob.text());
-  };
-
-  const stream = b64toStream(txt);
-  const resp = await decompressStream(stream);
-  return responseToJSON(resp);
-};
 
 const loadCaas = async (a) => {
   const encodedConfig = a.href.split('#')[1];
